+++ packages/catalog/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Muse Spark 1.3 Contributor now supports `max` on the Meta Model API and Muse Code providers ([#11788](https://github.com/can1357/oh-my-pi/issues/11788), [#12199](https://github.com/can1357/oh-my-pi/pull/12199) by [@mgpai22](https://github.com/mgpai22)).
+
 ## [18.2.1] - 2026-09-15
 
 ### Added